Finished move to proxy/models
