+++ python/ray/rllib/agents/sac/proxy/models/tf/modelv1_compat.py
@@ -3,7 +3,8 @@ import logging
 import numpy as np
 import tensorflow as tf
 
-from ray.rllib.agents.sac.sac.rllib_proxy._tf_model_v2 import ModelV2, TFModelV2
+from ray.rllib.agents.sac.proxy.models.model_v2 import ModelV2
+from ray.rllib.agents.sac.proxy.models.tf.tf_model_v2 import TFModelV2
 from ray.rllib.agents.sac.proxy.utils.tf_ops import scope_vars
 from ray.rllib.models.misc import normc_initializer, linear
 from ray.rllib.utils.annotations import override
