+++ python/ray/rllib/agents/sac/proxy/policy/dynamic_tf_policy.py
@@ -8,9 +8,7 @@ import tensorflow as tf
 from ray.rllib.agents.sac.proxy.policy.tf_policy import TFPolicy
 
 from ray.rllib.agents.sac.proxy.utils.tracking_dict import UsageTrackingDict
-from ray.rllib.agents.sac.sac.rllib_proxy._needs_patches import (
-    ModelCatalog
-)
+from ray.rllib.agents.sac.proxy.models.catalog import ModelCatalog
 from ray.rllib.evaluation.policy_graph import PolicyGraph as Policy
 from ray.rllib.evaluation.sample_batch import SampleBatch
 from ray.rllib.utils.annotations import override
