+++ python/ray/rllib/agents/sac/sac_model.py
@@ -1,8 +1,8 @@
 import numpy as np
 import tensorflow as tf
 import tensorflow_probability as tfp
+from ray.rllib.agents.sac.proxy.models.tf.tf_model_v2 import TFModelV2
 from ray.rllib.agents.sac.sac.custom_model import build_fcn
-from ray.rllib.agents.sac.sac.rllib_proxy._tf_model_v2 import TFModelV2
 from ray.rllib.utils.annotations import override
